+++ src/test/java/com/framework/helper/waitHelper.java
@@ -1,0 +1,4 @@
+package com.framework.helper;
+
+public class waitHelper {
+}
